refactor(chat): split ChatDialog panels into components
